Moving useFilters to filter directory in components
